# logging: reuse LoggingCaptureHandler instance since it's expensive to create