Previously, a LoggingCaptureHandler was instantiated for each test's setup/call/teardown which turns out to be expensive.

Instead, only keep one instance and reset it between runs.

(Previously in #7214. Now I'm pretty sure it will interact well with the other PRs).

<details>
<summary>Benchmark</summary>

```py
import pytest
@pytest.mark.parametrize("x", range(5000))
def test_foo(x): pass
```

Before:
```
============================ 5000 passed in 11.87s =============================
         10866808 function calls (10313462 primitive calls) in 12.163 seconds
```

After:

```
============================ 5000 passed in 11.25s =============================
         10431870 function calls (9878523 primitive calls) in 11.541 seconds
````
</details>