@bluetech 2020-05-18
Thanks for the review @thisch.

> It would have been helpful for the review, if you mentioned why you changed the catch_log related code in the commit msg.

The reason is that there is no longer a different handler for each `_runtest_for`, so the previous solution of keeping references to the handlers of each phase no longer works. So instead, now it only keeps a reference to the *records* of each phase, which get replaces each phase and stay in tact.